requirements.txt
discord.py
flask
pytz
apscheduler